kb_api/app/main/project/requirements.txt
aiofiles==0.4.0
aiopg==0.15.0
aioredis==1.1.0
amqp==2.3.2
async-timeout==3.0.1
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
gunicorn==19.9.0
hiredis==0.2.0
httptools==0.0.11
kombu==4.2.1
multidict==4.4.2
psycopg2==2.7.5
pycparser==2.19
PyJWT==1.6.4
pytz==2018.6
redis==2.10.6
sanic==0.8.3
six==1.11.0
ujson==1.35
uvloop==0.11.2
vine==1.1.4
websockets==5.0.1
